Improve theory behind anchors and links
Summary:

Right now documentation.js uses a blend of different approaches to
linking and needs a unified, tested system. The main change that this PR
makes so far is to avoid ever using 'slugifying' on namespaces in HTML
mode. Before we were doing so for headings, but not subheadings, and for
inline links but not links in the themes.

There's still research to be done here to figure out whether this is
robust enough - location hashes can include # and . characters, but can
they include everything else necessary in slugs? And does lodash's `<%-`
interpolation do enough to protect us from invalid HTML?

We'll still need to use slugifying for Markdown, since GitHub will
always create its own slugs.

* Fixes #525

TODO:

* Make anchors for sub-items linkable again. Main items have that nice
  hoverable link icon, but individual items don't.