fix: @name tag disconnects comment (#897)
The previous support for @name tags suppressing inference from
surrounding code was ad-hoc and incomplete. This commit makes a @name
tag (or a tag acting as a @kind/@name shorthand combo, like @function)
detach the `ast` field from the comment context, so that the surrounding
code will be universally ignored per the JSDoc spec.