feat(inferTypes): Support class properties
Previously we did not properly infer params on methods that were located on class properties. This
changes things so that we do. It carefully tries to avoid messing up the existing support of class
type annotations, which are similar to class properties at the AST level, but do not have an
associated node, only a type alias.

Fixes #1043