Remove unmaintained dependency: disparity
The dependency disparity has not been updated for 4 years and is using
a very old version of the diff package which has a security alert.

This alert is not relevant to the current usage of the package, however
it does generate noise when installing. Additionally, it seems unwise to
be relying on unmaintained packages.

This commit replaces disparity with diff and adds some code to color the
output which is based on disparity's logic.